+++ crates/ui/src/input/element.rs
@@ -514,10 +514,10 @@ impl Element for TextElement {
 
         // Set Root focused_input when self is focused
         if focused {
-            let input_view = self.input.clone();
-            if Root::read(window, cx).focused_input.as_ref() != Some(&input_view) {
+            let state = self.input.clone();
+            if Root::read(window, cx).focused_input.as_ref() != Some(&state) {
                 Root::update(window, cx, |root, _, cx| {
-                    root.focused_input = Some(input_view);
+                    root.focused_input = Some(state);
                     cx.notify();
                 });
             }
@@ -525,9 +525,9 @@ impl Element for TextElement {
 
         // And reset focused_input when next_frame start
         window.on_next_frame({
-            let input_view = self.input.clone();
+            let state = self.input.clone();
             move |window, cx| {
-                if !focused && Root::read(window, cx).focused_input.as_ref() == Some(&input_view) {
+                if !focused && Root::read(window, cx).focused_input.as_ref() == Some(&state) {
                     Root::update(window, cx, |root, _, cx| {
                         root.focused_input = None;
                         cx.notify();
