+++ crates/ui/src/input/mod.rs
@@ -6,9 +6,11 @@ mod input;
 mod mask_pattern;
 mod number_input;
 mod otp_input;
+mod text_input;
 
 pub(crate) use clear_button::*;
 pub use input::*;
 pub use mask_pattern::MaskPattern;
 pub use number_input::{NumberInput, NumberInputEvent, StepAction};
 pub use otp_input::*;
+pub use text_input::*;
